CDAP-14827 Added admin call to get namespace information
Added a method to get information about a namespace through the
Admin. Also added a generation field to namespaces that can
be used by consumers to know when a namespace has been deleted
and then recreated.

This feature will be used in dataprep to include the namespace
generation in entity row keys to avoid showing entities from
a previously existing namespace that was deleted and then
recreated.