CDAP-14840 add system app api
Introduced a new cdap-system-app-api module and corresponding
cdap-system-app-unit-test module for unit tests.

The new system-app-api module contains additional configurers
and contexts that are available to CDAP Services. These
classes allow applications to create system tables and use
them at runtime. Only applications deployed in the system
namespace are allowed to use the new classes. Attempts to
deploy an app in a user namespace will result in deployment
failure.

As part of this work, the restriction that system artifacts
cannot be added through the REST API was removed. This is because
unit tests require a way to add artifacts to the system namespace,
and also because the restriction is no long required now that
artifact authorization exists.